Move response passing test
